add: independent widget for the appbar content
